Move `OperationFields` into own module
